# aspnetcore/src/IO.Swagger/Controllers/DSSConstraintsApi.cs
/*
 * UTM DSS/USS API
 *
 * Interface to Discovery and Synchronization Service (DSS) and UAS Service Suppliers (USS) used by participating clients to discover and interoperate.  Unless otherwise specified, fields specified in a message but not declared in the API shall be ignored.
 *
 * OpenAPI spec version: 1.0.0-oas3
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;
using IO.Swagger.Security;
using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class DSSConstraintsApiController : ControllerBase
    { 
        /// <summary>
        /// /dss/constraints
        /// </summary>
        /// <remarks>Retrieve all Constraints in the given area during the given time.  Note that references to some Constraints lying entirely outside the requested area may be returned as well.</remarks>
        /// <param name="area">The area in which to search for Constraints.  References to some Constraints near this area but wholly outside it may also be returned.</param>
        /// <param name="timeStart">If specified, indicates non-interest in any Constraints that end before this time.  RFC 3339 format, per OpenAPI specification.</param>
        /// <param name="timeEnd">If specified, indicates non-interest in any Constraints that start after this time.  RFC 3339 format, per OpenAPI specification.</param>
        /// <response code="200">Constraints were successfully retrieved.</response>
        /// <response code="400">One or more input parameters were missing or invalid.</response>
        /// <response code="401">Bearer access token was not provided in Authorization header, token could not be decoded, or token was invalid.</response>
        /// <response code="403">The access token was decoded successfully but did not include a scope appropriate to this endpoint.</response>
        /// <response code="413">The requested area was too large.</response>
        [HttpGet]
        [Route("/v1/dss/constraints")]
        [ValidateModelState]
        [SwaggerOperation("V1DssConstraintsGet")]
        [SwaggerResponse(statusCode: 200, type: typeof(SearchConstraintReferencesResponse), description: "Constraints were successfully retrieved.")]
        [SwaggerResponse(statusCode: 400, type: typeof(ErrorResponse), description: "One or more input parameters were missing or invalid.")]
        [SwaggerResponse(statusCode: 401, type: typeof(ErrorResponse), description: "Bearer access token was not provided in Authorization header, token could not be decoded, or token was invalid.")]
        [SwaggerResponse(statusCode: 403, type: typeof(ErrorResponse), description: "The access token was decoded successfully but did not include a scope appropriate to this endpoint.")]
        [SwaggerResponse(statusCode: 413, type: typeof(ErrorResponse), description: "The requested area was too large.")]
        public virtual IActionResult V1DssConstraintsGet([FromQuery][Required()]Polygon area, [FromQuery][Required()]Time timeStart, [FromQuery][Required()]Time timeEnd)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(SearchConstraintReferencesResponse));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400, default(ErrorResponse));

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401, default(ErrorResponse));

            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403, default(ErrorResponse));

            //TODO: Uncomment the next line to return response 413 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(413, default(ErrorResponse));
            string exampleJson = null;
            exampleJson = "{\n  \"constraint_references\" : [ {\n    \"owner\" : \"uss_owner\",\n    \"feedback_url\" : \"https://my_utm_uss.com/feedback\",\n    \"time_start\" : \"\",\n    \"ovn\" : \"\",\n    \"id\" : \"\",\n    \"time_end\" : \"\",\n    \"details_url\" : \"https://my_utm_uss.com/constraints\"\n  }, {\n    \"owner\" : \"uss_owner\",\n    \"feedback_url\" : \"https://my_utm_uss.com/feedback\",\n    \"time_start\" : \"\",\n    \"ovn\" : \"\",\n    \"id\" : \"\",\n    \"time_end\" : \"\",\n    \"details_url\" : \"https://my_utm_uss.com/constraints\"\n  } ]\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<SearchConstraintReferencesResponse>(exampleJson)
                        : default(SearchConstraintReferencesResponse);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// /dss/constraints/{id}
        /// </summary>
        /// <remarks>Delete a Constraint.</remarks>
        /// <param name="id">EntityUUID of the Constraint.</param>
        /// <response code="200">The Constraint was successfully removed from the DSS.</response>
        /// <response code="400">* One or more input parameters were missing or invalid. * The request attempted to mutate the Constraint in a disallowed way.</response>
        /// <response code="401">Bearer access token was not provided in Authorization header, token could not be decoded, or token was invalid.</response>
        /// <response code="403">The access token was decoded successfully but did not include a scope appropriate to this endpoint.</response>
        /// <response code="409">* A Constraint with the specified ID already exists. * Despite repeated attempts, the DSS was unable to complete the update because of other simultaneous changes.</response>
        [HttpDelete]
        [Route("/v1/dss/constraints/{id}")]
        [ValidateModelState]
        [SwaggerOperation("V1DssConstraintsIdDelete")]
        [SwaggerResponse(statusCode: 200, type: typeof(ChangeConstraintReferenceResponse), description: "The Constraint was successfully removed from the DSS.")]
        [SwaggerResponse(statusCode: 400, type: typeof(ErrorResponse), description: "* One or more input parameters were missing or invalid. * The request attempted to mutate the Constraint in a disallowed way.")]
        [SwaggerResponse(statusCode: 401, type: typeof(ErrorResponse), description: "Bearer access token was not provided in Authorization header, token could not be decoded, or token was invalid.")]
        [SwaggerResponse(statusCode: 403, type: typeof(ErrorResponse), description: "The access token was decoded successfully but did not include a scope appropriate to this endpoint.")]
        [SwaggerResponse(statusCode: 409, type: typeof(ErrorResponse), description: "* A Constraint with the specified ID already exists. * Despite repeated attempts, the DSS was unable to complete the update because of other simultaneous changes.")]
        public virtual IActionResult V1DssConstraintsIdDelete([FromRoute][Required]EntityUUID id)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(ChangeConstraintReferenceResponse));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400, default(ErrorResponse));

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401, default(ErrorResponse));

            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403, default(ErrorResponse));

            //TODO: Uncomment the next line to return response 409 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(409, default(ErrorResponse));
            string exampleJson = null;
            exampleJson = "{\n  \"constraint_reference\" : \"\",\n  \"subscribers\" : [ {\n    \"subscriptions\" : [ {\n      \"subscription_id\" : \"\",\n      \"notification_index\" : 0\n    }, {\n      \"subscription_id\" : \"\",\n      \"notification_index\" : 0\n    } ],\n    \"url\" : \"\"\n  }, {\n    \"subscriptions\" : [ {\n      \"subscription_id\" : \"\",\n      \"notification_index\" : 0\n    }, {\n      \"subscription_id\" : \"\",\n      \"notification_index\" : 0\n    } ],\n    \"url\" : \"\"\n  } ]\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<ChangeConstraintReferenceResponse>(exampleJson)
                        : default(ChangeConstraintReferenceResponse);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// /dss/constraints/{id}
        /// </summary>
        /// <remarks>Retrieve Constraint (reference) information from the DSS.</remarks>
        /// <param name="id">EntityUUID of the Constraint.</param>
        /// <response code="200">Constraint was retrieved successfully.</response>
        /// <response code="400">One or more input parameters were missing or invalid.</response>
        /// <response code="401">Bearer access token was not provided in Authorization header, token could not be decoded, or token was invalid.</response>
        /// <response code="403">The access token was decoded successfully but did not include a scope appropriate to this endpoint.</response>
        /// <response code="404">The requested Entity could not be found.</response>
        [HttpGet]
        [Route("/v1/dss/constraints/{id}")]
        [ValidateModelState]
        [SwaggerOperation("V1DssConstraintsIdGet")]
        [SwaggerResponse(statusCode: 200, type: typeof(GetConstraintReferenceResponse), description: "Constraint was retrieved successfully.")]
        [SwaggerResponse(statusCode: 400, type: typeof(ErrorResponse), description: "One or more input parameters were missing or invalid.")]
        [SwaggerResponse(statusCode: 401, type: typeof(ErrorResponse), description: "Bearer access token was not provided in Authorization header, token could not be decoded, or token was invalid.")]
        [SwaggerResponse(statusCode: 403, type: typeof(ErrorResponse), description: "The access token was decoded successfully but did not include a scope appropriate to this endpoint.")]
        [SwaggerResponse(statusCode: 404, type: typeof(ErrorResponse), description: "The requested Entity could not be found.")]
        public virtual IActionResult V1DssConstraintsIdGet([FromRoute][Required]EntityUUID id)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(GetConstraintReferenceResponse));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400, default(ErrorResponse));

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401, default(ErrorResponse));

            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403, default(ErrorResponse));

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404, default(ErrorResponse));
            string exampleJson = null;
            exampleJson = "{\n  \"constraint_reference\" : {\n    \"owner\" : \"uss_owner\",\n    \"feedback_url\" : \"https://my_utm_uss.com/feedback\",\n    \"time_start\" : \"\",\n    \"ovn\" : \"\",\n    \"id\" : \"\",\n    \"time_end\" : \"\",\n    \"details_url\" : \"https://my_utm_uss.com/constraints\"\n  }\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<GetConstraintReferenceResponse>(exampleJson)
                        : default(GetConstraintReferenceResponse);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <param name="body"></param>
        /// <param name="id">EntityUUID of the Constraint.</param>
        /// <response code="200">A Constraint was created successfully in the DSS.</response>
        /// <response code="400">* One or more input parameters were missing or invalid. * The request attempted to mutate the Constraint in a disallowed way.</response>
        /// <response code="401">Bearer access token was not provided in Authorization header, token could not be decoded, or token was invalid.</response>
        /// <response code="403">The access token was decoded successfully but did not include a scope appropriate to this endpoint.</response>
        /// <response code="409">* A Constraint with the specified ID already exists. * Despite repeated attempts, the DSS was unable to complete the update because of other simultaneous changes.</response>
        /// <response code="413">The area of the Operation is too large.</response>
        [HttpPut]
        [Route("/v1/dss/constraints/{id}")]
        [ValidateModelState]
        [SwaggerOperation("V1DssConstraintsIdPut")]
        [SwaggerResponse(statusCode: 200, type: typeof(ChangeConstraintReferenceResponse), description: "A Constraint was created successfully in the DSS.")]
        [SwaggerResponse(statusCode: 400, type: typeof(ErrorResponse), description: "* One or more input parameters were missing or invalid. * The request attempted to mutate the Constraint in a disallowed way.")]
        [SwaggerResponse(statusCode: 401, type: typeof(ErrorResponse), description: "Bearer access token was not provided in Authorization header, token could not be decoded, or token was invalid.")]
        [SwaggerResponse(statusCode: 403, type: typeof(ErrorResponse), description: "The access token was decoded successfully but did not include a scope appropriate to this endpoint.")]
        [SwaggerResponse(statusCode: 409, type: typeof(ErrorResponse), description: "* A Constraint with the specified ID already exists. * Despite repeated attempts, the DSS was unable to complete the update because of other simultaneous changes.")]
        [SwaggerResponse(statusCode: 413, type: typeof(ErrorResponse), description: "The area of the Operation is too large.")]
        public virtual IActionResult V1DssConstraintsIdPut([FromBody]CreateConstraintReferenceParameters body, [FromRoute][Required]EntityUUID id)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(ChangeConstraintReferenceResponse));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400, default(ErrorResponse));

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401, default(ErrorResponse));

            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403, default(ErrorResponse));

            //TODO: Uncomment the next line to return response 409 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(409, default(ErrorResponse));

            //TODO: Uncomment the next line to return response 413 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(413, default(ErrorResponse));
            string exampleJson = null;
            exampleJson = "{\n  \"constraint_reference\" : \"\",\n  \"subscribers\" : [ {\n    \"subscriptions\" : [ {\n      \"subscription_id\" : \"\",\n      \"notification_index\" : 0\n    }, {\n      \"subscription_id\" : \"\",\n      \"notification_index\" : 0\n    } ],\n    \"url\" : \"\"\n  }, {\n    \"subscriptions\" : [ {\n      \"subscription_id\" : \"\",\n      \"notification_index\" : 0\n    }, {\n      \"subscription_id\" : \"\",\n      \"notification_index\" : 0\n    } ],\n    \"url\" : \"\"\n  } ]\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<ChangeConstraintReferenceResponse>(exampleJson)
                        : default(ChangeConstraintReferenceResponse);            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}