# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.2",
  "info" : {
    "title" : "UTM DSS/USS API",
    "description" : "Interface to Discovery and Synchronization Service (DSS) and UAS Service Suppliers (USS) used by participating clients to discover and interoperate.\n\nUnless otherwise specified, fields specified in a message but not declared in the API shall be ignored.",
    "version" : "1.0.0-oas3"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "security" : [ {
    "Authority" : [ "utm.read.operations", "utm.write.operations", "utm.read.constraints", "utm.write.constraints", "utm.read.subscriptions", "utm.write.subscriptions" ]
  } ],
...
  "paths" : {
...
    "/v1/dss/constraints" : {
      "summary" : "Query all Constraints in the DSS that intersect with a specified 4D-Volume.",
      "get" : {
        "tags" : [ "DSS Constraints" ],
        "summary" : "/dss/constraints",
        "description" : "Retrieve all Constraints in the given area during the given time.  Note that references to some Constraints lying entirely outside the requested area may be returned as well.",
        "parameters" : [ {
          "name" : "area",
          "in" : "query",
          "description" : "The area in which to search for Constraints.  References to some Constraints near this area but wholly outside it may also be returned.",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Polygon"
          }
        }, {
          "name" : "time_start",
          "in" : "query",
          "description" : "If specified, indicates non-interest in any Constraints that end before this time.  RFC 3339 format, per OpenAPI specification.",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Time"
          }
        }, {
          "name" : "time_end",
          "in" : "query",
          "description" : "If specified, indicates non-interest in any Constraints that start after this time.  RFC 3339 format, per OpenAPI specification.",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Time"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Constraints were successfully retrieved.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SearchConstraintReferencesResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "One or more input parameters were missing or invalid.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Bearer access token was not provided in Authorization header, token could not be decoded, or token was invalid.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The access token was decoded successfully but did not include a scope appropriate to this endpoint.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413" : {
            "description" : "The requested area was too large.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "Authority" : [ "utm.read.constraints" ]
        } ]
      }
    },
    "/v1/dss/constraints/{id}" : {
      "summary" : "Create/Update/Query a specific Constraint in the DSS.",
      "get" : {
        "tags" : [ "DSS Constraints" ],
        "summary" : "/dss/constraints/{id}",
        "description" : "Retrieve Constraint (reference) information from the DSS.",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "EntityUUID of the Constraint.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/EntityUUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Constraint was retrieved successfully.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GetConstraintReferenceResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "One or more input parameters were missing or invalid.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Bearer access token was not provided in Authorization header, token could not be decoded, or token was invalid.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The access token was decoded successfully but did not include a scope appropriate to this endpoint.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "The requested Entity could not be found.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "Authority" : [ "utm.read.constraints" ]
        } ]
      },
      "put" : {
        "tags" : [ "DSS Constraints" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "EntityUUID of the Constraint.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/EntityUUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateConstraintReferenceParameters"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "A Constraint was created successfully in the DSS.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ChangeConstraintReferenceResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "* One or more input parameters were missing or invalid.\n* The request attempted to mutate the Constraint in a disallowed way.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Bearer access token was not provided in Authorization header, token could not be decoded, or token was invalid.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The access token was decoded successfully but did not include a scope appropriate to this endpoint.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "* A Constraint with the specified ID already exists.\n* Despite repeated attempts, the DSS was unable to complete the update because of other simultaneous changes.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413" : {
            "description" : "The area of the Operation is too large.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "DSS Constraints" ],
        "summary" : "/dss/constraints/{id}",
        "description" : "Delete a Constraint.",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "EntityUUID of the Constraint.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/EntityUUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The Constraint was successfully removed from the DSS.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ChangeConstraintReferenceResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "* One or more input parameters were missing or invalid.\n* The request attempted to mutate the Constraint in a disallowed way.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Bearer access token was not provided in Authorization header, token could not be decoded, or token was invalid.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The access token was decoded successfully but did not include a scope appropriate to this endpoint.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "* A Constraint with the specified ID already exists.\n* Despite repeated attempts, the DSS was unable to complete the update because of other simultaneous changes.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "Authority" : [ "utm.write.constraints" ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Time" : {
        "type" : "string",
        "format" : "date-time"
      },
...
      "UUIDv4" : {
        "maxLength" : 36,
        "minLength" : 36,
        "type" : "string",
        "description" : "UUID v4.",
        "format" : "uuid",
        "example" : "03e5572a-f733-49af-bc14-8a18bd53ee39"
      },
...
      "EntityUUID" : {
        "description" : "Universally-unique identifier for an Entity communicated through the DSS.  Formatted as UUIDv4.",
        "anyOf" : [ {
          "$ref" : "#/components/schemas/UUIDv4"
        } ]
      },
...
      "ErrorResponse" : {
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string",
            "description" : "Human-readable message indicating what error occurred and/or why."
          }
        },
        "description" : "Data provided when an off-nominal condition was encountered."
      },
...
      "LatLngPoint" : {
        "required" : [ "lat", "lng" ],
        "type" : "object",
        "properties" : {
          "lng" : {
            "$ref" : "#/components/schemas/Longitude"
          },
          "lat" : {
            "$ref" : "#/components/schemas/Latitude"
          }
        },
        "description" : "Point on the earth's surface."
      },
...
      "Polygon" : {
        "required" : [ "vertices" ],
        "type" : "object",
        "properties" : {
          "vertices" : {
            "minItems" : 3,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/LatLngPoint"
            }
          }
        },
        "description" : "An enclosed area on the earth.\nThe bounding edges of this polygon shall be the shortest paths between connected vertices.  This means, for instance, that the edge between two points both defined at a particular latitude is not generally contained at that latitude.\nThe winding order shall be interpreted as the order which produces the smaller area.\nThe path between two vertices shall be the shortest possible path between those vertices.\nEdges may not cross.\nVertices may not be duplicated.  In particular, the final polygon vertex shall not be identical to the first vertex."
      },
...
      "GetConstraintReferenceResponse" : {
        "required" : [ "constraint_reference" ],
        "type" : "object",
        "properties" : {
          "constraint_reference" : {
            "$ref" : "#/components/schemas/ConstraintReference"
          }
        },
        "description" : "Response to DSS request for the ConstraintReference with the given ID.",
        "example" : {
          "constraint_reference" : {
            "owner" : "uss_owner",
            "feedback_url" : "https://my_utm_uss.com/feedback",
            "time_start" : "",
            "ovn" : "",
            "id" : "",
            "time_end" : "",
            "details_url" : "https://my_utm_uss.com/constraints"
          }
        }
      },
...
      "SearchConstraintReferencesResponse" : {
        "required" : [ "constraint_references" ],
        "type" : "object",
        "properties" : {
          "constraint_references" : {
            "type" : "array",
            "description" : "ConstraintReferences in the area of interest.",
            "items" : {
              "$ref" : "#/components/schemas/ConstraintReference"
            }
          }
        },
        "description" : "Response to DSS query for ConstraintReferences in an area of interest.",
        "example" : {
          "constraint_references" : [ {
            "owner" : "uss_owner",
            "feedback_url" : "https://my_utm_uss.com/feedback",
            "time_start" : "",
            "ovn" : "",
            "id" : "",
            "time_end" : "",
            "details_url" : "https://my_utm_uss.com/constraints"
          }, {
            "owner" : "uss_owner",
            "feedback_url" : "https://my_utm_uss.com/feedback",
            "time_start" : "",
            "ovn" : "",
            "id" : "",
            "time_end" : "",
            "details_url" : "https://my_utm_uss.com/constraints"
          } ]
        }
      },
...
    },
    "securitySchemes" : {
      "Authority" : {
        "type" : "oauth2",
        "description" : "Authorization from, or on behalf of, an authorization authority.  This authority shall issue access tokens that are JSON Web Tokens as defined in RFC 7519, using the `RS256` algorithm for the signature, and publish to all providers the public key for verifying that signature.\n\nThe following fields shall be included in the JWT claim for access tokens issued by this authority:\n* `iss`, with the URL at which the token generation request was received.\n* `exp`, with a time no further than 1 hour in the future.\n* `sub`, with unique ID of the client requesting the access token.\n* `scope`, with an array of strings indicating the scopes granted.\n\nMultiple scopes for a single token shall be granted if requested.  The tokens granted by this authority shall protect against reuse of received tokens to impersonate the sender to other recipients (via use of the `aud` claim or other means).\n\nClients shall provide these access tokens in an `Authorization` header in the form `Bearer <token>` in accordance with RFC 6750.",
        "flows" : {
          "clientCredentials" : {
            "tokenUrl" : "https://token_vendor.com/oauth/token",
            "scopes" : {
              "utm.read.operations" : "Client may query, read, or create subscriptions for Operations for UTM.",
              "utm.write.operations" : "Client may create, edit, and delete Operations for UTM.",
              "utm.read.constraints" : "Client may query, read, or create subscriptions for Constraints for UTM.",
              "utm.write.constraints" : "Client may create, edit, and delete Constraints for UTM."
            }
          }
        }
      }
    }
  }
}